modified custom bits section to fix EEPROM read bug
In the stepsFromDistance function, There was a EEPROM.read function to pull the custom bits heights, where it should have been EEPROM.get.   Because the address was written into EEPROM with a put, a read would not pull back binary.   Updated and fixed bug of the height of the bit was not being added to the bit travel on automated centering of the custom bits